backend/requirements.txt
fastapi
uvicorn
motor
pymongo
openai
anthropic
google-genai
google-generativeai
google-auth
requests
datasets
scikit-learn
sentence-transformers
numpy
pandas
python-dotenv
